Added redis clearing shutdown logic, added graceful SIGTERM for MM
